… CMakeLists.txt
… win32/Makefile.a64