The ld1{4 reg} variant saves us instructions
and only adds 3 cycles of latency to load 3
more neon/asimd registers worth of data.

Before:
```
 Tool: minigzip   Levels: 0-9       
 Runs: 100        Trim worst: 40        

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.008%      0.040/0.083/0.102/0.014        0.099/0.143/0.162/0.017      211,973,953
 1     44.414%      2.875/2.919/2.945/0.020        0.930/1.004/1.040/0.025       94,139,819
 2     35.521%      4.949/5.072/5.108/0.030        0.977/1.037/1.068/0.023       75,289,481
 3     33.882%      6.610/6.683/6.715/0.026        0.913/0.997/1.039/0.033       71,815,574
 4     33.175%      7.811/7.881/7.915/0.025        0.907/0.968/0.996/0.021       70,317,604
 5     32.661%      8.361/8.439/8.471/0.024        0.864/0.949/0.984/0.028       69,227,371
 6     32.507%   10.167/10.259/10.291/0.028        0.887/0.951/0.979/0.021       68,902,008
 7     32.255%   14.870/14.949/14.982/0.026        0.840/0.931/0.967/0.030       68,366,928
 8     32.167%   21.945/21.996/22.032/0.023        0.853/0.936/0.959/0.023       68,180,702
 9     31.887%   22.258/22.306/22.338/0.021        0.872/0.951/0.982/0.022       67,586,442

 avg1  40.848%                       10.059                          0.887
 avg2  45.386%                       11.176                          0.985
 tot                               6035.267                        532.020      865,799,882
```

After:

```
 Tool: minigzip   Levels: 0-9       
 Runs: 100        Trim worst: 40        

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.008%      0.020/0.081/0.108/0.019        0.091/0.143/0.168/0.018      211,973,953
 1     44.414%      2.782/2.879/2.914/0.027        0.940/1.000/1.028/0.019       94,139,819
 2     35.521%      4.987/5.037/5.069/0.021        0.958/1.032/1.067/0.025       75,289,481
 3     33.882%      6.522/6.631/6.664/0.032        0.906/0.992/1.027/0.032       71,815,574
 4     33.175%      7.690/7.835/7.866/0.031        0.897/0.957/0.993/0.026       70,317,604
 5     32.661%      8.313/8.386/8.420/0.027        0.866/0.948/0.976/0.022       69,227,371
 6     32.507%   10.144/10.217/10.253/0.027        0.865/0.936/0.973/0.025       68,902,008
 7     32.255%   14.853/14.904/14.931/0.020        0.873/0.933/0.964/0.022       68,366,928
 8     32.167%   21.899/21.957/21.992/0.023        0.877/0.936/0.962/0.019       68,180,702
 9     31.887%   22.193/22.278/22.321/0.029        0.887/0.952/0.984/0.026       67,586,442

 avg1  40.848%                       10.021                          0.883
 avg2  45.386%                       11.134                          0.981
 tot                               6012.416                        529.747      865,799,882
```