@KungFuJesus 2021-12-14
Tested this on an odroid n2