@KungFuJesus 2021-12-15
Ugh, yet another intrinsic not supported until way too late with GCC.  I guess we can do a compiler check to see if this is supported and fall back to single loads?