@nmoinvaz 2021-12-15
Maybe just ifdef check for older GCC version?