@KungFuJesus 2021-12-15
> Maybe just ifdef check for older GCC version?

Yeah I probably could but that might leave clang behind.  Given Apple's newest direction, we should probably generically detect this intrinsic pre-build.  Seems silly that we have to, but when gcc 9.3.x is too old I think it might be necessary.