@mtl1979 2021-12-15
Configure time check is what we have done before for most optional features... Benefit is that we don't have to do separate checks for gcc and clang as already mentioned. For some devices gcc 6.5 is still the latest usable version due to incompatibility issues in gcc 7.x series and later. For same reason Ubuntu 16.04 is latest usable version for some devices and 18.04 for others.