@KungFuJesus 2022-01-13
I'll probably get to this before too long, but it'll probably give us better resolution if I wait until after the benchmark PR goes through, since I can independently assess the functions that are doing this 64 byte load independently.